requirements.txt
biopython==1.81
scikit-learn==1.2.2
networkx==3.1